+++ config/src/config.rs
@@ -649,9 +649,6 @@ pub struct Config {
 
     #[dynamic(default)]
     pub quote_dropped_files: DroppedFileQuoting,
-
-    #[dynamic(default = "default_focus_change_repaint_delay")]
-    pub x11_focus_change_repaint_delay_ms: u64,
 }
 impl_lua_conversion_dynamic!(Config);
 
@@ -1349,10 +1346,6 @@ fn default_inactive_pane_hsb() -> HsbTransform {
         saturation: 0.9,
         hue: 1.0,
     }
-}
-
-fn default_focus_change_repaint_delay() -> u64 {
-    100
 }
 
 #[derive(FromDynamic, ToDynamic, Clone, Copy, Debug)]